back/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
distlib==0.3.6
Django==4.1.7
django-crispy-forms==2.0
django-environ==0.10.0
filelock==3.10.7
flatbuffers==23.3.3
gast==0.4.0
gevent==22.10.2
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.53.0
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.1.0
imutils==0.5.4
jax==0.4.7
keras==2.12.0
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
pipenv==2023.3.20
platformdirs==3.2.0
protobuf==4.22.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyWavelets==1.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
sqlparse==0.4.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-cpu-aws==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tifffile==2023.3.21
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
zope.event==4.6
zope.interface==6.0